Build double-precision GROMACS >= 2023 w/ AVX512
Under StdEnv/2020 and with GCC + MKL, GROMACS 2019 and newer
were producing broken executables when compiled with AVX512
and double precision, therefore logic falls back to using AVX2.
This was confirmed to be the case for GROMACS 2019 and 2020.

A test under StdEnv/2023 with GROMACS 2023.3 suggests the issue
is fixed.

At a later date we should narrow in whether this was is repoducible
for GROMACS 2019, ... 2022 under StdEnv/2023 and whether it is
fixed for 2023 under StdEnv/2020 (GCC 9.3.0) as well.